Finalize favourites feature, upgrade gradle plugin

+++ app/src/main/java/com/example/kittens/data/network/models/SuccessResponse.kt
@@ -1,3 +1,3 @@
 package com.example.kittens.data.network.models
 
-data class SuccessResponse (val message: String, val id: Int?)
+data class SuccessResponse (val message: String, val id: Long?)
